Introduce a compositing_traits crate.
