+++ components/compositing/compositor.rs
@@ -56,6 +56,8 @@ use webrender;
 use webrender_traits::{self, ScrollEventPhase};
 use windowing::{self, MouseWindowEvent, WindowEvent, WindowMethods, WindowNavigateMsg};
 
+pub use compositing_traits::CompositingReason;
+
 #[derive(Debug, PartialEq)]
 enum UnableToComposite {
     NoContext,
@@ -2562,27 +2564,4 @@ impl<Window> CompositorEventListener for IOCompositor<Window> where Window: Wind
             warn!("Failed to send pipeline title ({}).", e);
         }
     }
-}
-
-/// Why we performed a composite. This is used for debugging.
-#[derive(Copy, Clone, PartialEq, Debug)]
-pub enum CompositingReason {
-    /// We hit the delayed composition timeout. (See `delayed_composition.rs`.)
-    DelayedCompositeTimeout,
-    /// The window has been scrolled and we're starting the first recomposite.
-    Scroll,
-    /// A scroll has continued and we need to recomposite again.
-    ContinueScroll,
-    /// We're performing the single composite in headless mode.
-    Headless,
-    /// We're performing a composite to run an animation.
-    Animation,
-    /// A new frame tree has been loaded.
-    NewFrameTree,
-    /// New painted buffers have been received.
-    NewPaintedBuffers,
-    /// The window has been zoomed.
-    Zoom,
-    /// A new WebRender frame has arrived.
-    NewWebRenderFrame,
 }
